+++ internal/amf/ngap/handle_pdu_session_resource_modify_response.go
@@ -19,7 +19,7 @@ func HandlePDUSessionResourceModifyResponse(ctx context.Context, amfInstance *am
 	}
 
 	ueConn.TouchLastSeen()
-	logger.WithTrace(ctx, ueConn.Log).Debug("Handle PDUSessionResourceModifyResponse", zap.Int64("AmfUeNgapID", ueConn.AmfUeNgapID))
+	logger.WithTrace(ctx, ueConn.Log).Debug("Handle PDUSessionResourceModifyResponse", zap.Int64("AmfUeNgapID", int64(ueConn.AmfUeNgapID)))
 
 	if msg.UserLocationInformation != nil {
 		ueConn.UpdateLocation(ctx, amfInstance, msg.UserLocationInformation)
